requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.9
django-allauth==0.34.0
django-appconf==1.0.2
django-braces==1.12.0
django-embed-video==1.1.2
django-environ==0.4.5
django-image-cropping==1.2.0
django-nested-admin==3.1.2
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.1.0
django-solo==1.1.3
django-suit-ckeditor==0.0.2
django-taggit==0.23.0
django-tinymce4-lite==1.7.2
djangorestframework==3.7.7
docopt==0.6.2
easy-thumbnails==2.5
html5lib==1.0.1
idna==2.6
jsmin==2.2.2
mercadopago==0.3.4
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
PyPDF2==1.26.0
python-monkey-business==1.0.0
python3-openid==3.1.0
pytz==2017.3
reportlab==3.5.9
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.15.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
urllib3==1.22
webencodings==0.5.1
xhtml2pdf==0.2.3
yarg==0.1.9
